requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
google-genai>=1.0.0
pydantic==2.6.3
supabase==2.3.5
